gen_ai/requirements.txt
crewai[tools]>=0.108.0
openai>=1.40.0
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn>=0.30.0
